functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
requests>=2.28
python-dotenv>=1.0
groq>=0.9.0
